caja-checkout

+++ database/migrations/2020_12_03_022819_create_tipo_habitacion_table.php
@@ -16,7 +16,7 @@ class CreateTipoHabitacionTable extends Migration
         Schema::create('tipohabitacion', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nombre', 20);
-            $table->integer('precio');
+            $table->decimal('precio', 10, 2);
             $table->timestamps();
         });
     }
